understading_python08
In this "Guess the Number" project, the user attempts to guess a randomly generated number between 1 and 100 within 10 attempts. The program validates user input to ensure it's a number within the specified range and provides feedback on each guess. If the user guesses correctly or exhausts their attempts, appropriate messages are displayed to conclude the game.